+++ src/main/java/com/example/fairshareapp/service/impl/SueldoServiceImpl.java
@@ -63,7 +63,7 @@ public class SueldoServiceImpl implements SueldoService {
 
         // Sincroniza el sueldo en el usuario para calculos proporcionales
         if (sueldoRequest.getMonto() != null) {
-            usuario.setSueldo(sueldoRequest.getMonto().doubleValue());
+            usuario.setSueldo(sueldoRequest.getMonto());
             usuarioRepository.save(usuario);
         }
 
@@ -126,7 +126,7 @@ public class SueldoServiceImpl implements SueldoService {
 
         if (sueldo.getUsuario() != null && sueldoRequest.getMonto() != null) {
             Usuario usuario = sueldo.getUsuario();
-            usuario.setSueldo(sueldoRequest.getMonto().doubleValue());
+            usuario.setSueldo(sueldoRequest.getMonto());
             usuarioRepository.save(usuario);
         }
